remove RequestUri and wire Uri through
